move ListSitesResponse to Models/Reponse/Site, add CustomDomainEntity, use it in SiteEntity. update SiteEntity display names
